requirements.txt
chardet==4.0.0
click==8.1.2
commonmark==0.9.1
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pygments==2.12.0
pygount==1.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
rich==12.3.0
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.1.1
zipp==3.8.0
